+++ src/Runtime/Optimized/Physics/IMmdWasmPhysicsRuntime.ts
@@ -17,16 +17,12 @@ export interface IMmdWasmPhysicsRuntime {
     /**
      * Physics max sub steps (default 5)
      *
-     * recommended value is 120 for high quality physics simulation
-     *
      * The maximum number of physics sub steps
      */
     maxSubSteps: number;
 
     /**
-     * Physics fixed time step (default 1 / 60)
-     *
-     * recommended value is 1 / 120 for high quality physics simulation
+     * Physics fixed time step (default 1 / 100)
      *
      * The fixed time step of physics
      */