+++ src/Runtime/Optimized/wasm_src/bullet_stdlib/stdio.cpp
@@ -1,6 +1,27 @@
 #include <stdio.h>
+#include <stdarg.h>
 
-// we don't need to implement printf
-
-void printf(const char* fmt, ...) {
+#ifdef BT_DEBUG
+extern "C" {
+    int bw_error(char const* message, va_list args);
 }
+#endif
+
+int printf(const char* fmt, ...) {
+#ifdef BT_DEBUG
+    va_list args;
+
+    va_start(args, fmt);
+    int bytes_written = bw_error(fmt, args);
+    va_end(args);
+
+    return bytes_written;
+#else
+    return 0;
+#endif
+}
+
+#ifdef BT_DEBUG
+void __debugbreak() {
+}
+#endif